backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
faster-whisper==0.10.0
ffmpeg-python==0.2.0
openai==1.10.0
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
pillow==10.2.0
numpy==1.26.3
